meepzorp/agents/task/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.26.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
requests==2.31.0
loguru==0.7.2 